% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conduct.MatrixTransformation_PGLS.R
\name{Conduct.MatrixTransformation_PGLS}
\alias{Conduct.MatrixTransformation_PGLS}
\title{Conduct.MatrixTransformation_PGLS: function to compute the phylogenetic transformation matrix that can be used to transform input trait values}
\usage{
Conduct.MatrixTransformation_PGLS(handle.Phylogeny)
}
\arguments{
\item{handle.Phylogeny}{Phylogeny object that is used to generate the matrix.P (phylogenetic transform matrix)}
}
\value{
matrix.P Matrix used for phylogenetic transformation of input trait data
}
\description{
This function returns a matrix that can be used to transform input trait values in phylogenetic space
}
\examples{

################
# load depends #
################
library(geiger)
library(ROBRT)

################################
# simulate tree and trait data #
################################
handle.Phylogeny <- sim.bdtree(b = 1, n = 10)
vector.TraitValue <- sim.char(handle.Phylogeny, par=1, model="BM", root=0)[,,1]

#####################
# transform to PGLS #
#####################
Conduct.MatrixTransformation_PGLS(handle.Phylogeny = handle.Phylogeny)
}
\keyword{Brownian}
\keyword{comparative}
\keyword{data}
\keyword{method,}
\keyword{motion,}
\keyword{phylogenetic}
\keyword{regression,}
\keyword{robust}
\keyword{trait}
