% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conduct.PhylogeneticRegression_PIC_M.R
\name{Conduct.PhylogeneticRegression_PIC_M}
\alias{Conduct.PhylogeneticRegression_PIC_M}
\title{Conduct.PhylogeneticRegression_PIC_M: function to conduct phylogenetic regression using PIC and M estimator}
\usage{
Conduct.PhylogeneticRegression_PIC_M(
  handle.Phylogeny,
  vector.Trait_X,
  vector.Trait_Y
)
}
\arguments{
\item{handle.Phylogeny}{Phylogeny object that is used to compute PICs}

\item{vector.Trait_X}{Vector containing values for independent (predictor) trait X}

\item{vector.Trait_Y}{Vector containing values for dependent (response) trait y}
}
\value{
handle.PIC_M Object containing results of M regression for PIC
}
\description{
This function returns a regression object that can be read using standard R functions (summary, lm, etc.)
}
\examples{

################
# load depends #
################
library(geiger)
library(ROBRT)
library(phytools)
library(MASS)
library(L1pack)

################################
# simulate tree and trait data #
################################
handle.Phylogeny <- sim.bdtree(b = 1, n = 10)
vector.TraitValue_X <- sim.char(handle.Phylogeny, par=1, model="BM", root=0)[,,1] # traits are independent 
vector.TraitValue_Y <- sim.char(handle.Phylogeny, par=1, model="BM", root=0)[,,1] # traits are independent 

#####################################
# Conduct M phylogenetic regression #
#####################################
summary(Conduct.PhylogeneticRegression_PIC_M(handle.Phylogeny = handle.Phylogeny, vector.Trait_X = vector.TraitValue_X, vector.Trait_Y = vector.TraitValue_Y))

handle.RESULTS_M <- Conduct.PhylogeneticRegression_PIC_M(handle.Phylogeny = handle.Phylogeny, vector.Trait_X = vector.TraitValue_X, vector.Trait_Y = vector.TraitValue_Y)
f.robftest(handle.RESULTS_M, var = "vector.PIC_X")$p.value
}
\keyword{Brownian}
\keyword{comparative}
\keyword{data}
\keyword{method,}
\keyword{motion,}
\keyword{phylogenetic}
\keyword{regression,}
\keyword{robust}
\keyword{trait}
