% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conduct.PhylogeneticRegression_PIC_S.R
\name{Conduct.PhylogeneticRegression_PIC_S}
\alias{Conduct.PhylogeneticRegression_PIC_S}
\title{Conduct.PhylogeneticRegression_PIC_S: function to conduct phylogenetic regression using PIC and S estimator}
\usage{
Conduct.PhylogeneticRegression_PIC_S(
  handle.Phylogeny,
  vector.Trait_X,
  vector.Trait_Y
)
}
\arguments{
\item{handle.Phylogeny}{Phylogeny object that is used to compute PICs}

\item{vector.Trait_X}{Vector containing values for independent (predictor) trait X}

\item{vector.Trait_Y}{Vector containing values for dependent (response) trait y}
}
\value{
handle.PIC_S Object containing results of S regression for PIC
}
\description{
This function returns a regression object that can be read using standard R functions (summary, lm, etc.)
}
\examples{

################
# load depends #
################
library(geiger)
library(ROBRT)
library(phytools)
library(MASS)
library(L1pack)
library(robust)

################################
# simulate tree and trait data #
################################
handle.Phylogeny <- sim.bdtree(b = 1, n = 10)
vector.TraitValue_X <- sim.char(handle.Phylogeny, par=1, model="BM", root=0)[,,1] # traits are independent 
vector.TraitValue_Y <- sim.char(handle.Phylogeny, par=1, model="BM", root=0)[,,1] # traits are independent 

######################################
# conduct S phylogenetic regression #
######################################
summary(Conduct.PhylogeneticRegression_PIC_S(handle.Phylogeny = handle.Phylogeny, vector.Trait_X = vector.TraitValue_X, vector.Trait_Y = vector.TraitValue_Y))
}
\keyword{Brownian}
\keyword{comparative}
\keyword{data}
\keyword{method,}
\keyword{motion,}
\keyword{phylogenetic}
\keyword{regression,}
\keyword{robust}
\keyword{trait}
