% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conduct.Robust_PhylogeneticRegression_PGLS.R
\name{Conduct.Robust_PhylogeneticRegression_PGLS}
\alias{Conduct.Robust_PhylogeneticRegression_PGLS}
\title{Conduct.Robust_PhylogeneticRegression_PGLS: function to conduct robust phylogenetic regression using L2, L1, S, M, and/or MM}
\usage{
Conduct.Robust_PhylogeneticRegression_PGLS(
  handle.Phylogeny,
  vector.Trait_X,
  vector.Trait_Y,
  vector.Estimators
)
}
\arguments{
\item{handle.Phylogeny}{Phylogeny object that is used to compute PGLS}

\item{vector.Trait_X}{Vector containing values for independent (predictor) trait X}

\item{vector.Trait_Y}{Vector containing values for dependent (response) trait y}

\item{vector.Estimators}{Vector where each element is a string that is "L1", "L2", "S", "M", or "MM"}
}
\value{
list.RESULTS List containing regression results for each estimator included in vector.Estimators
}
\description{
This function returns a list of robust phylogenetic regression results for L2, L1, S, M, and/or MM
}
\examples{

}
\keyword{Brownian}
\keyword{comparative}
\keyword{data}
\keyword{method,}
\keyword{motion,}
\keyword{phylogenetic}
\keyword{regression,}
\keyword{robust}
\keyword{trait}
