% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conduct.Robust_PhylogeneticRegression_PIC.R
\name{Conduct.Robust_PhylogeneticRegression_PIC}
\alias{Conduct.Robust_PhylogeneticRegression_PIC}
\title{Conduct.Robust_PhylogeneticRegression_PIC: function to conduct robust phylogenetic regression using L2, L1, S, M, and/or MM}
\usage{
Conduct.Robust_PhylogeneticRegression_PIC(
  handle.Phylogeny,
  vector.Trait_X,
  vector.Trait_Y,
  vector.Estimators
)
}
\arguments{
\item{handle.Phylogeny}{Phylogeny object that is used to compute PICs}

\item{vector.Trait_X}{Vector containing values for independent (predictor) trait X}

\item{vector.Trait_Y}{Vector containing values for dependent (response) trait y}

\item{vector.Estimators}{Vector where each element is a string that is "L1", "L2", "S", "M", or "MM"}
}
\value{
list.RESULTS List containing regression results for each estimator included in vector.Estimators
}
\description{
This function returns a list of robust phylogenetic regression results for L2, L1, S, M, and/or MM
}
\examples{

################
# load depends #
################
library(geiger)
library(ROBRT)
library(phytools)
library(MASS)
library(L1pack)
library(robust)

################################
# simulate tree and trait data #
################################
handle.Phylogeny <- sim.bdtree(b = 1, n = 10)
vector.TraitValue_X <- sim.char(handle.Phylogeny, par=1, model="BM", root=0)[,,1] # traits are independent 
vector.TraitValue_Y <- sim.char(handle.Phylogeny, par=1, model="BM", root=0)[,,1] # traits are independent 
vector.Estimators <- c("L2", "L1", "M", "MM", "S")

######################################
# conduct S phylogenetic regression #
######################################
handle.RESULTS_PIC <- Conduct.Robust_PhylogeneticRegression_PIC(handle.Phylogeny = handle.Phylogeny, 
                                                                vector.Trait_X = vector.TraitValue_X, 
                                                                vector.Trait_Y = vector.TraitValue_Y, 
                                                                vector.Estimators = vector.Estimators)
}
\keyword{Brownian}
\keyword{comparative}
\keyword{data}
\keyword{method,}
\keyword{motion,}
\keyword{phylogenetic}
\keyword{regression,}
\keyword{robust}
\keyword{trait}
